refactor: rename permit2 files
