+++ src/components/Swap/SwapActionButton/AllowanceButton.tsx
@@ -3,7 +3,7 @@ import { PERMIT2_ADDRESS } from '@uniswap/permit2-sdk'
 import ActionButton from 'components/ActionButton'
 import EtherscanLink from 'components/EtherscanLink'
 import { usePendingApproval } from 'hooks/transactions'
-import { AllowanceRequired } from 'hooks/usePermit2'
+import { AllowanceRequired } from 'hooks/usePermit2Allowance'
 import { Spinner } from 'icons'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { ApprovalTransactionInfo } from 'state/transactions'
